+++ projects/demo-maps/src/app/route-components/route-example-layout/route-map5.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, HostBinding, OnDestroy } from '@angular/core';
 import { LayersService, RasterLayer, VectorLayer, LayerGroup, Layer, WmtsLayer } from '@dlr-eoc/services-layers';
 import { MapStateService } from '@dlr-eoc/services-map-state';
-import { osm, esri_world_imagery, esri_ocean_imagery, eoc_litemap, esri_grey_canvas, esri_nav_charts, open_sea_map } from '@dlr-eoc/base-layers-raster';
+import { EocLitemap, OsmTileLayer, OpenSeaMap, EocLiteoverlayTile, WorldReliefBwTile } from '@dlr-eoc/base-layers-raster';
 import { MapOlService } from '@dlr-eoc/map-ol';
 import { Subscription } from 'rxjs';
 
@@ -17,7 +17,7 @@ export class RouteMap5Component implements OnInit, OnDestroy {
   controls: { attribution?: boolean, scaleLine?: boolean, zoom?: boolean, crosshair?: boolean };
 
 
-  geojsonData: any;
+  public geojsonData: any;
   showTable = false;
   subs: Subscription[] = [];
   constructor(
@@ -41,31 +41,12 @@ export class RouteMap5Component implements OnInit, OnDestroy {
   }
 
   addBaseLayers() {
-    const eocLitemapLayer = new eoc_litemap({
-      legendImg: null,
-      id: 'eoc_litemap_base',
+    const eocLitemapLayer = new EocLitemap({
       visible: true
     });
 
     // not working in WGS84
-    const worldRelief = new WmtsLayer({
-      type: 'wmts',
-      url: 'https://tiles.geoservice.dlr.de/service/wmts',
-      name: 'Relief',
-      id: 'world_relief_bw',
-      params: {
-        layer: 'eoc:world_relief_bw',
-        style: '_empty',
-        matrixSetOptions: {
-          matrixSet: 'EPSG:3857',
-          tileMatrixPrefix: 'EPSG:3857'
-        }
-      },
-      visible: false,
-      description: 'eoc:world_relief_bw as web map tile service',
-      attribution: 'Relief: <a src="https://www.dlr.de/eoc">DLR/EOC</>',
-      legendImg: ''
-    });
+    const worldRelief = new WorldReliefBwTile();
 
     const layers = [eocLitemapLayer, worldRelief];
 
@@ -74,41 +55,6 @@ export class RouteMap5Component implements OnInit, OnDestroy {
   }
 
   addLayers() {
-    const gufLayer = new RasterLayer({
-      type: 'wms',
-      url: 'https://geoservice.dlr.de/eoc/land/wms',
-      name: 'GUF Mosaic',
-      id: 'GUF28_DLR_v1_Mosaic',
-      params: {
-        layers: 'GUF28_DLR_v1_Mosaic',
-        styles: 'guf_8bit'
-      },
-      visible: false,
-      description: 'GUF28_DLR_v1_Mosaic',
-      attribution: ' | GUF®: <a href="https://www.dlr.de/eoc/en/desktopdefault.aspx/tabid-9628/16557_read-40454/">DLR License</a>',
-      legendImg: ''
-    });
-
-    const TDM90DEMLayer = new WmtsLayer({
-      type: 'wmts',
-      url: 'https://tiles.geoservice.dlr.de/service/wmts',
-      name: 'TDM90 DEM',
-      id: 'TDM90_DEM',
-      params: {
-        layer: 'TDM90_DEM',
-        style: 'default',
-        matrixSetOptions: {
-          matrixSet: 'EPSG:3857',
-          tileMatrixPrefix: 'EPSG:3857'
-        },
-        format: 'image/png'
-      },
-      visible: false,
-      description: 'TDM90_DEM',
-      attribution: ' | TDM90 Data ©: <a href="http://www.dlr.de" target="_blank">DLR</a>  licensed for <a rel="license" target="_blank" href="https://geoservice.dlr.de/resources/licenses/tdm90/License_for_the_Utilization_of_90m_DEM_for_Scientific_Use.pdf">scientific use</a>',
-      legendImg: ''
-    });
-
     this.geojsonData = {
       type: 'FeatureCollection',
       features: [
@@ -226,96 +172,18 @@ export class RouteMap5Component implements OnInit, OnDestroy {
       popup: true
     });
 
-    const vectorLayer2 = new VectorLayer({
-      id: 'geojson_test_2',
-      name: 'Vector Layer in Group',
-      type: 'geojson',
-      data: {
-        type: 'FeatureCollection',
-        features: [
-          {
-            type: 'Feature',
-            properties: { title: 'Rectangle' },
-            geometry: {
-              type: 'Polygon',
-              coordinates: [
-                [
-                  [
-                    10.986328125,
-                    43.89789239125797
-                  ],
-                  [
-                    11.755371093749998,
-                    43.89789239125797
-                  ],
-                  [
-                    11.755371093749998,
-                    44.41808794374846
-                  ],
-                  [
-                    10.986328125,
-                    44.41808794374846
-                  ],
-                  [
-                    10.986328125,
-                    43.89789239125797
-                  ]
-                ]
-              ]
-            }
-          }
-        ]
-      },
-      visible: false,
-      popup: true
-    });
-
-    const esriImageLayer = new esri_world_imagery();
-
-    const esriGreyLayer = new esri_grey_canvas({
-      removable: true,
-      legendImg: null,
-    });
-
-    const esriOceanImageryLayer = new esri_ocean_imagery({
-      removable: true,
-      legendImg: null,
-      id: 'esri_ocean_base'
-    });
-
-    const osmLayer = new osm();
-    /** add a Group of layers */
-
-
-    const groupLayer = new LayerGroup({
-      id: 'group_1',
-      name: 'Test Group',
-      layers: [esriOceanImageryLayer, osmLayer, esriImageLayer]
-    });
-
-    const groupLayer2 = new LayerGroup({
-      id: 'group_2',
-      name: 'Test Group 2',
-      layers: [TDM90DEMLayer, vectorLayer2, esriGreyLayer]
-    });
-
-    const overlays = [gufLayer, groupLayer2, vectorLayer, groupLayer];
-    overlays.map(layer => {
-      if (layer instanceof Layer) {
-        this.layersSvc.addLayer(layer, 'Layers');
-      } else if (layer instanceof LayerGroup) {
-        this.layersSvc.addLayerGroup(layer);
-      }
-    });
+    this.layersSvc.addLayer(vectorLayer, 'Layers');
   }
 
   checkVisibleLayers() {
     const sub = this.layersSvc.getLayers().subscribe((layers) => {
       layers.map((layer) => {
-        if (layer.id === 'geojson_test' && layer.visible === true) {
-          this.showTable = true;
-        } else if (layer.id === 'geojson_test' && layer.visible === false) {
-          this.showTable = false;
+        if (layer) {
+          if (layer.id === 'geojson_test' && layer.visible === true) {
+            this.showTable = true;
+          } else if (layer.id === 'geojson_test' && layer.visible === false) {
+            this.showTable = false;
+          }
         }
       });
     });
@@ -323,10 +191,8 @@ export class RouteMap5Component implements OnInit, OnDestroy {
   }
 
   addOverlays() {
-    const layerOnToppOfAll = new esri_nav_charts();
-    const openSeaMapOnTopp = new open_sea_map();
-    this.layersSvc.addLayer(layerOnToppOfAll, 'Overlays');
-    this.layersSvc.addLayer(openSeaMapOnTopp, 'Overlays');
+    const onTopp = new EocLiteoverlayTile();
+    this.layersSvc.addLayer(onTopp, 'Overlays');
   }
 
   ngOnDestroy() {